moved functions over, need new commit file from josh to import
